split route components in another file
